feat(ui,web): move wiki domain components into @repowise/ui/wiki
Six presentational components: code-block, confidence-badge,
git-history-panel, mermaid-diagram, table-of-contents, wiki-markdown.

git-history-panel swaps GitMetadataResponse for the canonical
GitMetadata from @repowise/types/git. Wiki components reach across
domains for git-side primitives (CommitCategorySparkline,
CoChangeList) via @repowise/ui/git/* — no fetch coupling. Adds
mermaid, react-markdown, and remark-gfm as @repowise/ui deps. The
Next.js RSC wiki-renderer (next-mdx-remote/rsc + shiki),
regenerate-button (mutation + GenerationProgress), version-history
(usePageVersions hook), and security-panel (SWR) stay in packages/web
for now.

Subpath export ./wiki/* added; wiki-renderer rewires its sibling
imports to @repowise/ui/wiki/*; consumers in app/page,
docs-viewer, wiki/[...slug]/page, and graph-doc-panel rewired.
