+++ packages/ui/package.json
@@ -22,6 +22,7 @@
     "./coverage": "./src/coverage/index.ts",
     "./coverage/*": "./src/coverage/*.tsx",
     "./wiki": "./src/wiki/index.ts",
+    "./wiki/*": "./src/wiki/*.tsx",
     "./chat": "./src/chat/index.ts",
     "./dashboard": "./src/dashboard/index.ts",
     "./workspace": "./src/workspace/index.ts",
@@ -58,7 +59,10 @@
     "d3-force": "^3.0.0",
     "d3-hierarchy": "^3.1.2",
     "lucide-react": "^1.7.0",
+    "mermaid": "^11.4.0",
+    "react-markdown": "^10.1.0",
     "recharts": "^3.8.1",
+    "remark-gfm": "^4.0.1",
     "tailwind-merge": "^3.5.0"
   },
   "peerDependencies": {
