+++ packages/web/src/app/repos/[id]/wiki/[...slug]/page.tsx
@@ -6,11 +6,11 @@ import { getRepo } from "@/lib/api/repos";
 import { getPageById, getPageVersions } from "@/lib/api/pages";
 import { getGitMetadata } from "@/lib/api/git";
 import { getGraphMetrics } from "@/lib/api/graph";
-import { ConfidenceBadge } from "@/components/wiki/confidence-badge";
+import { ConfidenceBadge } from "@repowise/ui/wiki/confidence-badge";
 import { WikiRenderer } from "@/components/wiki/wiki-renderer";
-import { TableOfContents } from "@/components/wiki/table-of-contents";
+import { TableOfContents } from "@repowise/ui/wiki/table-of-contents";
 import { RegenerateButton } from "@/components/wiki/regenerate-button";
-import { GitHistoryPanel } from "@/components/wiki/git-history-panel";
+import { GitHistoryPanel } from "@repowise/ui/wiki/git-history-panel";
 import { SecurityPanel } from "@/components/wiki/security-panel";
 import { Badge } from "@repowise/ui/ui/badge";
 import { Separator } from "@repowise/ui/ui/separator";
