+++ packages/web/src/components/graph/graph-doc-panel.tsx
@@ -15,7 +15,7 @@ import { cn } from "@/lib/utils/cn";
 import { Button } from "@repowise/ui/ui/button";
 import { Badge } from "@repowise/ui/ui/badge";
 import { ChatMarkdown } from "@/components/chat/chat-markdown";
-import { ConfidenceBadge } from "@/components/wiki/confidence-badge";
+import { ConfidenceBadge } from "@repowise/ui/wiki/confidence-badge";
 import { usePage } from "@/lib/hooks/use-page";
 import { formatRelativeTime } from "@repowise/ui/lib/format";
